src/job_search_gui/requirements.txt
pytest-mock
pytest==8.3.4
Markdown~=3.7
reportlab~=4.3.1
google-generativeai
pylint
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pylint